+++ srcs/libft/libft.h
@@ -44,7 +44,7 @@ int				ft_strncmp(const char *s, const char *s2, size_t n);
 size_t			ft_strlcpy(char *dst, const char *src, size_t dstsize);
 size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
 char			*ft_strnstr(const char *string, const char *needle, size_t len);
-int				ft_atoi(const char *str);
+ssize_t				ft_atoi(const char *str);
 void			*ft_calloc(size_t count, size_t size);
 char			*ft_strdup(const char	*s1);
 char			*ft_substr(char const *s, unsigned int start, size_t len);
